// frontend/ReceiptTracker/src/app/api/model/tokenOut.ts
/**
 * Receipt Tracker API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TokenOut { 
    access_token: string;
    refresh_token: string;
    token_type?: string;
}